add unit test

+++ app/src/main/java/com/example/weather/model/Wind.kt
@@ -3,7 +3,7 @@ import com.google.gson.annotations.SerializedName
 data class Wind(
 
 	@field:SerializedName("deg")
-	val deg: Int? = null,
+	val deg: Double? = null,
 
 	@field:SerializedName("speed")
 	val speed: String? = null
